cruds sale consignment

+++ app/models/Sim_sale_consignments_model.php
@@ -7,15 +7,18 @@ class Sim_sale_consignments_model extends CI_Model
         parent::__construct();
     }
 
-    function GetAutocomplete($options = array())
+    function GetAutocomplete($term)
     {
-
-    $this->db->select('name');
-    $this->db->like('name', $options['name'], 'after');
-
-     
-     $query = $this->db->get('sim_groups');
-     return $query->result();
+        $this->db->select('id, name', FALSE);
+        $this->db->where("{$this->db->dbprefix('sim_groups')}.name LIKE '%" . $term . "%");
+        
+        $q = $this->db->get('sim_groups');
+        if ($q->num_rows() > 0) {
+            foreach (($q->result()) as $row) {
+                $data[] = $row;
+            }
+            return $data;
+        }
     }
 
     public function getBranches()
@@ -32,6 +35,17 @@ class Sim_sale_consignments_model extends CI_Model
 
     public function getShops(){
     	$q = $this->db->get("sim_shops");
+        if ($q->num_rows() > 0) {
+            foreach (($q->result()) as $row) {
+                $data[] = $row;
+            }
+            return $data;
+        }
+        return FALSE;
+    }
+
+     public function getGroups(){
+        $q = $this->db->get("sim_groups");
         if ($q->num_rows() > 0) {
             foreach (($q->result()) as $row) {
                 $data[] = $row;
@@ -91,24 +105,46 @@ class Sim_sale_consignments_model extends CI_Model
 
     public function deleteSimGroupFromSlCongt($gId)
     {
+        $this->db->where(array('use_group_id' => $gId,'is_sale'=> 1));
+        $simsInGroup = $this->db->get("sale_consignment_detail");
+        if($simsInGroup->num_rows() > 0){
+            return FALSE;
+        }
+
         if ($this->db->delete('sale_consignment_detail', array('use_group_id' => $gId))) {
             return true;
         }
         return FALSE;
     }
 
+    public function deleteSaleCognt($id)
+    {
+         $this->db->select("sale_consignment_detail.*")
+            ->where(array('is_sale'=> 1, 'use_sale_consignment_id'=> $id));
+        $q =  $this->db->get('sale_consignment_detail');
+
+        if($q->num_rows() > 0){
+            return FALSE;
+        }
+
+        if ($this->db->delete('sale_consignment_detail', array('use_sale_consignment_id' => $id))) {
+            if($this->db->delete('sim_sale_consignments', array('id' => $id))){
+                return true;
+            }
+        }
+        return FALSE;
+    }
+
     public function getSaleConsignents($id){
-        $this->db->select("sale_consignment_detail.use_group_id as id, sim_sale_consignments.date_consign, sim_groups.name, sim_shops.shop, CONCAT(".$this->db->dbprefix('sim_locations').".name, ' (', ".$this->db->dbprefix('sim_branches').".branch_name, ')') as locationName, sim_branches.facebook_name, users.username, sim_sale_consignments.reference_note")
-            ->join('sim_groups', 'sim_groups.id = sale_consignment_detail.use_group_id')
-            ->join('sim_sale_consignments', 'sim_sale_consignments.id = sale_consignment_detail.use_sale_consignment_id')
+       $this->db->select($this->db->dbprefix('sim_sale_consignments') . ".id as id, sim_sale_consignments.date_consign, sim_shops.shop, CONCAT(".$this->db->dbprefix('sim_locations').".name, ' (', ".$this->db->dbprefix('sim_branches').".branch_name, ')') as locationName, sim_branches.facebook_name, users.username, sim_sale_consignments.reference_note")
+            ->join('sale_consignment_detail', 'sale_consignment_detail.use_sale_consignment_id = sim_sale_consignments.id')
             ->join('sim_branches', 'sim_branches.id = sim_sale_consignments.use_sim_branches_id')
             ->join('sim_shops', 'sim_shops.id = sim_branches.use_shop_id')
             ->join('sim_locations', 'sim_locations.id = sim_branches.use_sim_location_id')
             ->join('users', 'users.id = sim_sale_consignments.use_sale_man_id')
-            ->where('users.id', $this->session->userdata('user_id'))
-            ->where('sale_consignment_detail.use_group_id', $id)
-            ->group_by('sale_consignment_detail.use_group_id', $id);
-        $q =  $this->db->get('sale_consignment_detail');
+            ->where(array('users.id' => $this->session->userdata('user_id'), 'sim_sale_consignments.id' => $id))
+            ->group_by('sale_consignment_detail.use_sale_consignment_id');
+        $q =  $this->db->get('sim_sale_consignments');
         if ($q->num_rows() > 0) {
             foreach (($q->result()) as $row) {
                 $data[] = $row;
@@ -151,5 +187,34 @@ class Sim_sale_consignments_model extends CI_Model
             return $data;
         }
         return FALSE;
+    }
+
+    public function addGroupToShop($saleConId, $gId){
+        $isGroupExist = $this->db->get("sale_consignment_detail", array('use_sim_group_id'=> $gId, 'use_sale_consignment_id' => $saleConId));
+        if($isGroupExist->num_rows() > 0){
+            return false;
+        }
+
+        $sims = $this->db->get("sim", array('use_sim_group_id'=> $gId));
+        if ($sims->num_rows() > 0) {
+            $success = true;
+            foreach (($sims->result()) as $row) {
+            $data = array(
+                'use_group_id' => $gId,
+                'use_sim_id' => $row->id,
+                'use_sale_consignment_id' => $saleConId,
+                'is_sale' => 0
+            );
+
+            $status = $this->db->insert("sale_consignment_detail", $data);
+            $success = $success && $status;
+            }
+
+            if($success){
+                return true;
+            }
+       }
+
+       return false;
     }
 }
